New look to the gui and custom JGauge Component

+++ rc-telemetry-app/src/main/java/com/franciscodadone/model/Accelerometer.java
@@ -1,6 +1,6 @@
 package com.franciscodadone.model;
 
 public class Accelerometer {
-    public static int z;
-    public static int maxZ;
+    public static float z;
+    public static float maxZ;
 }
